ISSUE #4911 - move vertical calibration helpers to actual file
