+++ frontend/src/v5/ui/routes/dashboard/projects/calibration/calibration.helpers.ts
@@ -15,6 +15,8 @@
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
+import { Coord2D, Transformation2D, Vector2D, Vector3D } from './calibration.types';
+
 export const UNITS_CONVERTION_FACTORS_TO_METRES = {
 	'm': 1,
 	'dm': 10,
@@ -30,3 +32,79 @@ export const getUnitsConvertionFactor = (drawingUnits, modelUnits) => {
 
 export const convertCoordUnits = (coord, convertionFactor: number) => coord?.map((point) => point * convertionFactor) || null;
 export const convertVectorUnits = (vector, convertionFactor: number) => vector.map((coord) => convertCoordUnits(coord, convertionFactor));
+
+export const getXYPlane = (vector: Vector3D) => vector.map((val) => [val[0], val[2]]) as Vector2D;
+
+export const addVectors = (vectorA: number[], vectorB: number[]) => {
+	const sumVector = [];
+	for (let i = 0; i < vectorA.length; i++) {
+		sumVector[i] = vectorB[i] + vectorA[i];
+	}
+	return sumVector;
+};
+
+export const subtractVectors = (vectorA: number[], vectorB: number[]) => {
+	const diffVector = [];
+	for (let i = 0; i < vectorA.length; i++) {
+		diffVector[i] = vectorA[i] - vectorB[i];
+	}
+	return diffVector;
+};
+
+const crossProduct = (matrixA: number[][], matrixB: number[][]) => {
+	const aNumRows = matrixA.length, aNumCols = matrixA[0].length,
+		bNumCols = matrixB[0].length,
+		result = new Array(aNumRows);
+	for (var r = 0; r < aNumRows; ++r) {
+		result[r] = new Array(bNumCols);
+		for (var c = 0; c < bNumCols; ++c) {
+			result[r][c] = 0;
+			for (var i = 0; i < aNumCols; ++i) {
+				result[r][c] += matrixA[r][i] * matrixB[i][c];
+			}
+		}
+	}
+	return result;
+};
+
+const dotProduct = (vectorA: number[], vectorB: number[]) => vectorA.reduce((acc, _, i) => acc + vectorA[i] * vectorB[i], 0);
+
+const getVectorMagnitude = (vector: number[]) => Math.sqrt(dotProduct(vector, vector));
+
+export const getTransformationMatrix = (vectorA, vectorB) => {
+	const diffA = subtractVectors(vectorA[1], vectorA[0]);
+	const diffB = subtractVectors(vectorB[1], vectorB[0]);
+	const magnitudeA = getVectorMagnitude(diffA);
+	const magnitudeB = getVectorMagnitude(diffB);
+	const scaleFactor = magnitudeB / magnitudeA;
+
+	// in order to know if angle is clockwise or anti-clockwise we find the cross product of both vectors and take the sign of the z-component
+	const crossProductZ = (diffA[0] * diffB[1]) - (diffA[1] * diffB[0]);
+	const directionFactor = crossProductZ < 0 ? 1 : -1;
+	const angle = Math.acos(dotProduct(diffA, diffB) / (magnitudeA * magnitudeB)) * directionFactor; // angle between vectors in radians
+
+	const scaleMatrix = [[ scaleFactor, 0], [0, scaleFactor]];
+	const rotationMatrix = [ // rotates 2D vector clockwise around origin by 'angle'
+		[Math.cos(angle), Math.sin(angle)],
+		[-Math.sin(angle), Math.cos(angle)],
+	];
+	const transformationMatrix = crossProduct(rotationMatrix, scaleMatrix) as Transformation2D; 
+	return transformationMatrix;
+};
+
+const transformVector = (v: number[], t: number[][]) => {
+	const newVector = [];
+	for (let row = 0; row < t.length; row++) {
+		let sum = 0;
+		for (let col = 0; col < t.length; col++) {
+			sum = sum + t[row][col] * v[col];
+		}
+		newVector.push(sum);
+	}
+	return newVector;
+};
+
+export const transformAndTranslate = (v: Coord2D, t: Transformation2D, offset: Coord2D) => {
+	const transformed = transformVector(v, t);
+	return addVectors(offset, transformed);
+};
